+++ CHANGELOG.md
@@ -4,11 +4,12 @@
 
 ### New Features
 - Added a display that shows the last time the hopper inside a minion has been emptied
-- Added a marker to the location of the last clicked minion for a couple of seconds (even through walls)
+- Added a marker to the last opened minion for a couple of seconds (seen through walls)
 - Added option to hide mob name tags close to minions
 - Added showing stars on all items (not only dungeon stars and master stars but also on crimson armours, cloaks and fishing rods)
 - Added a display timer that shows the real time
 - Added overlay features to the RNG meter inventory (highlight selected drop and floors without a drop and show floor)
+- Added minion hopper coins per day display (Using the held coins in the hopper and the last time the hopper was collected to calculate the coins a hopper collects in a day)
 
 ### Minor Changes
 - Summoning souls display is rendering better close to corners